+++ includes/ui/diagnostics.php
@@ -10,12 +10,11 @@ defined( '\\ABSPATH' ) || exit;
 global $wp_object_cache;
 
 $info = [];
-$dropins = [];
-$dropin = $this->validate_object_cache_dropin();
+$dropin = $plugin->validate_object_cache_dropin();
 $disabled = defined( 'WP_REDIS_DISABLED' ) && WP_REDIS_DISABLED;
 
-$info['Status'] = $this->get_status();
-$info['Client'] = $this->get_redis_client_name();
+$info['Status'] = $plugin->get_status();
+$info['Client'] = $plugin->get_redis_client_name();
 $info['Drop-in'] = $dropin ? 'Valid' : 'Invalid';
 $info['Disabled'] = $disabled ? 'Yes' : 'No';
 
@@ -45,7 +44,7 @@ if ( defined( 'HHVM_VERSION' ) ) {
     $info['HHVM Version'] = HHVM_VERSION;
 }
 
-$info['Redis Version'] = $this->get_redis_version() ?: 'Unknown';
+$info['Redis Version'] = $plugin->get_redis_version() ?: 'Unknown';
 
 $info['Multisite'] = is_multisite() ? 'Yes' : 'No';
 
@@ -118,21 +117,21 @@ if ( $dropin ) {
     );
 }
 
-foreach ( $info as $name => $value ) {
-    // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-    echo "{$name}: {$value}\r\n";
-}
+$dropins = [];
 
 foreach ( get_dropins() as $file => $details ) {
-    $dropins[ $file ] = sprintf(
-        ' - %s v%s by %s',
-        $details['Name'],
-        $details['Version'],
-        $details['Author']
-    );
+    $dropins[ $file ] = sprintf( '%s v%s by %s', $details['Name'], $details['Version'], $details['Author'] );
 }
 
-if ( ! empty( $dropins ) ) {
-    // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-    echo "Drop-ins: \r\n", implode( "\r\n", $dropins ), "\r\n";
+$info['Drop-ins'] = wp_json_encode(
+    array_values( $dropins ),
+    JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE
+);
+
+foreach ( $info as $name => $value ) {
+    if ( defined( 'WP_CLI' ) && WP_CLI ) {
+        WP_CLI::line( "{$name}: $value" );
+    } else {
+        echo "{$name}: {$value}\r\n";
+    }
 }