+++ includes/ui/tabs/diagnostics.php
@@ -7,6 +7,8 @@
 
 defined( '\\ABSPATH' ) || exit;
 
+$plugin = $this;
+
 ?>
 
 <p>